todolist test e controller modificados

+++ backend/src/controllers/TodoList.controller.js
@@ -42,6 +42,10 @@ module.exports = {
     async show(req, res) {
         let { id } = req.params;
 
+        if (isNaN(id)) {
+            return res.status(400).json("Envie um id válido no próximo request!");
+        }
+
         try {
             const todoResults = await db.query(
                 'SELECT * FROM todolists WHERE id = ($1)',
@@ -71,6 +75,10 @@ module.exports = {
         let { id } = req.params;
         let { title } = req.body;
 
+        if (isNaN(id)) {
+            return res.status(400).json("Envie um id válido no próximo request!");
+        }
+
         if (!title) {
             return res.status(400).json('Envie um título válido no próximo request');
         }
@@ -98,6 +106,10 @@ module.exports = {
     async delete(req, res) {
         try {
             let { id } = req.params;
+
+            if (isNaN(id)) {
+                return res.status(400).json("Envie um id válido no próximo request!");
+            }
 
             const result = await db.query(
                 'DELETE FROM todolists WHERE id = ($1)',
